Fix the rendered selection that is shifted by 1px (#9029)
The commit fixes an issue where the rendered selection is shifted by 1px. The
issue is easily reproducible by enabling the NestedHeaders plugin but it turned
out that the issue is more general.

The PR#7281 changed the API the element's widths is called from. It contributed
to the selection shift issue. Most of the DOM's API returns decimal values,
Handsontable and Walkontable depend on that values. Changing the `.offsetWidth`
to `getBoundingClientRect` creates this misalignment issue. In the cases when
the browser calculates the width as a decimal point with a fractional component
value (e.q 35.33px) the `.offsetWidth` returns 35 (rounded down <.49, rounded
up >.5) and the previous implementation returned 36 (rounded up).

As far as I know, the original issue #7262 was about fixing the AutoColumnSize
plugin. So I reverted the changes to the `ofsetWidth` helper and applied the
`getBoundingClientRect` to the GhostTable class only.