Fix an issue where the nested headers are not rendered (#9023)
The commit fixes an issue where the nested headers are not rendered in some
configurations. No matter if the `columns` or `startCols` options are used the
plugin always renders columns to the passed dataset columns length limit. I
believe it's not intended. The mentioned options allow virtually extending the
grid to a specific size and the plugin should respect that.