Fix severe performance issue when removing events
There is a performance bug in EventManager#clearEvents, where every
individual event being removed causes a linear search through all
existing event listeners.

Obviously, this scales really poorly. Example:

A 20x20 table of 400 cells, with up to 4 borders per cell will produce
1600 listeners, not including the myriad of other, unrelated listeners
that are potentially present in a table.

Listener removal iterates over all listeners, but listener removal
itself is also invoked for every single listener. This means that
removing the listener for a single border might result in 2560000 inner
iterations for this example.

It therefore seems that this formula gives a worst-case number of inner
iterations for a certain number `n` of listeners to remove:

    f(n) => n*n + f(n-1)

With our example of 1600 this results in roughly 1.4e9 inner iterations,
which takes roughly a minute to execute in my test environment.

This fix only does it once per border instead, which comes out to this
formula:

    f(n) => n + f(n-1)

These are just back-of-the-envelope calculations, but they match up with
the measured real-world numbers in my local tests:

Iterations of the event removal loop before the fix:
    Setting borders:       7997414 iterations
    Removing borders:   4292962614 iterations

Iterations of the event removal loop after the fix:
    Setting borders:          2838 iterations
    Removing borders:      2279707 iterations

This is still quite a significant number of iterations, but at least it
seems fast enough in practice so that a complete re-engineering of the
whole event architecture isn't strictly required.

The added test verifies that clearing multiple events never delegates to
EventManager#removeEventListener to prevent accidental performance
regressions.