Fix autocomplete suggest list for older version of FF (#9077)
The commit fixes an issue where the suggestion list of the autocomplete cell
editors is not updated correctly. I think that the issue is reproducible only
on the older versions of Firefox like 86, 88 (on Windows). On the v91 and
newer, I'm not able to recreate the issue (or it is more difficult to induce).

The commit adds a slight delay to the `queryChoices` method. Thanks to the
delay the value taken from the textarea element is the freshest value and the
list is updated always based on the full sentence typed into cell editor.