Refactoring of walkontable (#9005)
The commit adds several steps of refactoring, which target to improve the
developer experience while developing to Walkontable. The primary assumption
was to loosen the relations between internal components. To accomplish that, I
introduce the Inversion of Control with Dependency Injection through
constructors wherever where it was able. That allows to introduction
one-directional relation tree (in most cases). However, some classes have
circular dependencies by referencing Walkontable (Core) class. This makes it
impossible without deeper changes to the unification direction of relations.
For that reason, I introduced Data Access Object, which allows us to get the
required components without accessing the Core class dynamically. This is not
the perfect approach, because DAO in assumptions should provide only raw data
(and I added // TODO where that is broken). On the other side, it allows for
loosening relations between components. So, when continuing the further
improvements, probably needs to break the backward compatibility because to
remove circular dependencies requires redesigning the responsibilities of each
class in Walkontable and splitting into many with a single responsibility
(fixing the SOLID principles here).

Following the topic of backward compatibility. During the refactoring keeping
the backward compatibility is essential. And good refactoring cannot break
that. That really knots the hands, thus I introduce a new layer which makes me
able to loosen this knot. I added the facade layer, which is the external API
of Walkontable. The facade cares to keep the same API as before refactoring.
That makes me able to change some method names and map access to them in the
facade. This is a really useful tool for changing internal relations. However,
the complexity of the facade layer grows with the wideness of internal changes.
During this refactoring, I went to the point where I figured that the wider
changes will occur with an enormous increasing the logic in the facade layer.
Then makes the whole Walkontable even more difficult to maintain. In that
situation, I recommend redesigning and rewriting the WOT lib and then exposing
an entirely new API. In my human opinion, this is the song of the future
because we have got many more important things to improve than the deepest
layer of HOT. For that reason, I finished the refactoring for now.

Additionally, there are:
* splitted core and clone walkontable's classes, which makes code cleanest and
easiest to debug.
* accessing the settings indirectly from the wtSettings service injected into
components.
* added facadeGetter which should be used only when we provide reference to
instance externally to Walkontable. Using internally is FORBIDDEN.

Co-authored-by: Krzysztof Budnik <571316+budnix@users.noreply.github.com>