[RTL] Fix Text editor positioning and sizing (#9117)
The commit fixes the positioning and sizing misalignment of the cell's Text
editor in RTL mode.

Moreover, the change removes legacy CSS code. It causes the double border to
appear and only for the second column? weird. In LTR mode the border does not
appear (I guess the JS code removes it while table rendering) but in RTL does.