[RTL] Custom Borders direction-less API (#9144)
The commit implements the new API for objects that declares the borders
styling. The `CustomBorders` plugin allows changing the left and right borders
using the "left"/"right" property names. That is confusing when the table is
run in the RTL document mode.

The new API changes those names, so the `left` property becomes `start` and
`right` becomes an `end`. The `start` property depends on the document layout
direction pointing to the left border in LTR and the right border in RTL.

The proposed changes are backward compatible. It simply means that after the
library upgrades the developers do not have to change anything in their apps.
The "left"/"right" properties will work as previously. With indication that it
would only work in LTR document mode.