[RTL] Internal API naming (#9176)
To make the code easier to understand (less confusing) especially in the
context of RTL the commit changes the internal API naming. Most of the APIs
that use "left"/"right" words can be confused in RTL. In that mode "left" often
means "right" and vice versa. The change changes the naming to
"inlineStart"/"inlineEnd" or "start"/"end" depending on the context.

1. Changes to the overlays' CSS classes
The CSS classes are changed. The commit adds new names such as
`ht_clone_top_inline_start_corner`, `ht_clone_inline_start`, and
`ht_clone_bottom_inline_start_corner` simultaneously leaving the old classes in
the DOM for backward compatibility.

The commit 7c75528 (tests 550c310).

2. Changes to the other CSS class names
The `innerBorderLeft` CSS class is replaced with `innerBorderInlineStart`. The
old class is left in the DOM for backward compatibility.

The commit a820796 (tests f549bd6).

3. Changes internal API access to the Walkontable instance
To emphasize that the Walkontable rendering engine is not a part of public API
the property name is changed from `hot.view.wt` to `hot.view._wt`.

The commit 1d3310c.

4. Changes internal API access to the overlays' instances
The changes (`hot.view._wt`):

* `.wtOverlays.topLeftCornerOverlay` =>
`.wtOverlays.topInlineStartCornerOverlay`;
* `.wtOverlays.leftOverlay` => `.wtOverlays.inlineStartOverlay`;
* `.wtOverlays.bottomleftCornerOverlay` =>
`.wtOverlays.bottomInlineStartCornerOverlay`.
The commit 7c75528.

5. Changes to the CellCoords and CellRange classes.
The commit updates the method names and implementation logic to follow the new
direction-less terminology. To support the methods that point to the specific
direction (e.g. `getBottomRightCorner` or `isNorthEastOf` etc.) the above
classes from now on expect the constructor argument isRtl that allows
distinguishing the logic within. For that methods, the "left" is always
organoleptically "left". However, in most cases, to support LTR and RTL the
following methods should be used:

`getTopStartCorner`;
`getTopEndCorner`;
`getBottomStartCorner`;
`getBottomEndCorner`;
`getOuterTopStartCorner`;
`getOuterTopEndCorner`;
`getOuterBottomEndCorner`;
`getOuterBottomStartCorner`.
The methods are part of the new API. Depending on the layout mode they are
returning left or right coordinates.

The commit 595892d, d876b2b (tests 91264f4).