[RTL] Abstract the editor's positioning logic (#9201)
The commit adds the `getEditedCellRect` method. The method reduces DRY (Text
and Select editors have the same implementation) and allows developers to
position their editors perfectly right above the edited cell, no matter if the
edited cell belongs to the main, top, left, or bottom overlay.

The method returns an object with the following properties:

* `top` The top position relative to the table viewport;
* `start` The left (or right in RTL) position relative to the table viewport;
* `width` The cell's current width;
* `maxWidth` The maximum cell's width after which the editor goes out of the
table viewport;
* `height` The cell's current height;
* `maxHeight` The maximum cell's height after which the editor goes out of the
table viewport

To calculate correct values for cells that are edited within overlays there is
needed to implement a new `getOverlayOffset` method. The method works similar
to the `getScrollPosition` method but the returned values are different
depending on the type of the scrollable element. When Handsontable knows its
size its overlays position is always 0. The overlays are stuck to the table's
edge and in that case, the `getScrollPosition` only matters. When Handsontable
is initialized without size the Window object becomes the scrollable element.
In that case, while scrolling the browser viewport the overlays only moves when
the table's viewport leaves the viewport of the browser - to keep the overlays
stuck to the browser's viewport edge. The scroll position from
`getScrollOffset` method wouldn't fit here as it calculates the value always
from left/top 0. The `getOverlayOffset` method does the job by returning the
scroll position minus parent offset.

One of the biggest pros of the method is the fact the complexity of the logic
behind the overlay positioning is drastically reduced (for example InlineStart
overlay or TopInlineStartCorner overlay).