Change behavior of the Cmd/Ctrl + Arrow keyboard shortcuts (#9375)
The commit changes the behavior of the Cmd/Ctrl + Arrow keyboard shortcuts.

To implement a new behavior, there is a need to implement a new mechanism
(called `captureCtrl`) that disables capturing and returning the state of the
pressed `Command/Meta` keys. That method helper is broadly used within the
table and works independently from the registered keyboard shortcuts. For some
shortcuts, that use the same modifier keys, it's desirable to capture the
`Control/Command` key to prevent messing with the logic that they are
executing. For example, preventing the contiguous selection that happens when
the `Control/Command` is pressed.

Co-authored-by: Wojciech Szymański <141330+wszymanski@users.noreply.github.com>