Remove hardcoded "htLeft" alignment from demos (#9420)
The commit updates all demos (from the examples package) to add support for
different settings for the layout direction option. From now on, all demos
should look just as good as with left-to-right direction and right-to-left.

The task shows a need to expose a private API to the public. The
methods `isRtl`, `isLtr`, and `getDirectionFactor` in the context of demos (and
probably in the context of the user's apps) turned out to be very useful.

Co-authored-by: Jakub Wiśniewski <jakub.wisniewski@handsontable.com>