Fix inline start overlay desync issue after cell editing (#9422)
The commit fixes the issue about the inline-start overlay scroll
desynchronization with the main table. The reason why it happened was the fact
that there was a bad timing between clearing the row height after cell edit and
recalculating it again before table render. There was a timing window where the
edited row height returned an `undefined` value, hence the rendering engine
wrongly calculated the scroll position for the overlay. By changing the hook
from `beforeChange` to `beforeChangeRender` I'm able to minimize that render
"window" where the row height is not available yet to the minimum and I think
that it's sufficient to fix the bug.