Fix type error for validation process within multiple HF sheets (#9476)
The commit fixes a bug that causes a type error exception for Handsontable with
HF integration. The error happened in cases when the HF was passed as an
external engine and the engine has more than 1 sheet added. The Formulas plugin
incorrectly assumed that all sheets have a binding with the Handsontable
instance, hence it tried to validate the not existing table.

The fix adds the condition that checks if under the sheet name the HoT instance
exists. If not, the validation process is skipped.