Fix maximum call stack size in indexMapper (#9490)
The commit fixes the RangeError (maximum call stack size exceeded) that happens
when the Handsontable uses more hidden rows and/or columns than ~6000 records.
I found that the error happens in two places.

* In the public `getFirstNotHiddenIndex` method of the IndexMapper class;
* In the private method `findVisibleCoordsInRowsRange` of the VisualSelection
class;

The methods used the complexity of O(n) which combined with recursion caused
the methods to explode. For setup where all 6000 rows are hidden calling the
API as `hot.getFirstNotHiddenIndex(0, 1)` causes about 6 thousand recursion
calls until the method throws the RangeError.

The commit is not a breaking change. The original method will stay in the
public API till the next major release. But for now, it's marked as
`@deprecated` and it prints the warning message when used. The new bug-free and
more simplified `getNearestNotHiddenIndex` method from now on taking a place of
the original method.