Fix the Walkontable viewport calculations for oversized rows and columns (#9542)
The commit fixes a bug where Handsontable was not able to render correctly
oversized rows and/or columns. By saying oversized, I mean the columns that are
wider than the table's viewport width and rows that are longer than the table's
viewport height.

To make it work, there needed to fix two problems.

1. The fix for the Viewport calculation. There was a bug that caused not to
render non-fully visible records. The fix adds conditions that check whether
only the partially visible records are rendered in the viewport, and if true,
then the full table's render is performed.
2. The fix for the Scroll module. To support scrolling between partially
visible records, the class had to be refactored. The module allows scrolling
the viewport to the cell and choosing to which part of the table's edge the
cell will be stuck. For oversized records, this is not possible. The oversized
cells are always stuck to the top-left (or top-right in RTL) position (the same
works in other common spreadsheets available in the market).