Fix scrollable container element stretching for mobiles (#9567)
The commit fixes the bug where the container of the scrollable element was
stretched. This causes the table to render more wide table viewport than
necessary in RTL mode. The bug was caused by the CSS `left: -10000px` property
that is used to "hide" mobile selection handlers for an editor. The fix is
about changing the property to `right: -10000px` - for RTL (using
`inset-inline-*` props). Thanks to that, the selection handlers are always
moved from the outside of the viewport in the opposite direction than the
scroll direction. Thus, no stretching viewport happens.