Move the `@angular/*` package to peer dependencies (#9575)
The commit moves the `@angular/*` packages to the peer dependencies. Updates
the `npm audit` command from `npm audit --production` to `npm audit --omit=dev
--omit=peer` - the previous argument is deprecated and add new examples for
Angular 9, 10, 11, 12, 13 and upgraded the latest Angular example to version
14.

The full list of the PR changes:

* Move the `@angular/*` dependencies to the peers;
* Add examples for Angular 9 (v9-lts), 10 (v10-lts), 11 (v11-lts), 12
(v12-lts), 13 (v13-lts) and 14 (as latest);
* Update npm audit command;
* Update puppeteer dependencies thanks to that, the examples work not only on
x86 CPU architecture but on the ARM (Apple M1, M2, etc.) as well;
* Speed up cleaning and installation process by adding tasks parallelization;
* Improve the Code Examples GitHub Workflow by adding a step that runs the
tests for each example;
* Fix Angular examples for versions 9, 12, and 13. There was a problem with the
app setup that won't work with a symlinked wrapper. Switching the
`preserveSymlinks` option to `false` fixed a problem.