Fix Angular wrapper peers deps and update examples (#9617)
The commit fixes the regression bug introduced within 12.1.0 by changing the
peers' dependencies from ^9.0.0 to >=9.0.0. Moreover, the PR synchronizes the
example dependencies with the specific Angular version and unifies the Angular
examples.

To make sure that the wrapper will build always with Angular 9 the packages
with the ^9.0.0 range are added to `devDependencies`.

The change temporarily downgrades the `@handsontable/angular` package in the
"next" examples. Otherwise, the example installation will fail. The change does
not influence the "next" examples' dependencies. The wrapper version is
overwritten while creating the new release.

And finally, code Angular examples for 12.1.0 are broken and cause tests to
fail so they are removed. They will re-appear after a new 12.1.1 release.

Co-authored-by: Marcin Warpechowski <warpech@gmail.com>