Change the Docs deployment architecture (#9651)
The commit changes the Docs code architecture and deployment by moving the
specific Docs versions into specific production branches and creating versioned
Docker images.

The commit contains the following more significant changes:

* Docs deployment
* Previously: Documentation deployment was triggered automatically by merging
the ./docs/** changes to the develop branch, no matter what Docs version was
changed.
* Now: The change has to be pushed to the prod-docs/{MAJOR}.{MINOR} branch. The
develop branch holds content for the next releases and can be manually deployed
to the staging.
* Generating Docs API
* Previously: Docs API was generated manually by executing npm run docs:api
command and generated files are pushed to the release branch (while code
freeze).
* Now: The API is generated automatically after the Handsontable release
process is finished. The files are generated in the release script and then
pushed to the newly created branch - if it is the MAJOR or MINOR release or,
merged to the latest branch's Docs version.
* Stateless Docs images issues
* After splitting the Docs version into separate Docker images it turned out
that there are some challenges in how to update the Docs versions dropdown
list, canonical URLs, and other dynamic variables without rebuilding the older
Docs versions. So, there an idea arose that the variables will be created and
saved into the JSON file. The file then is consumed by the older Docs images so
that way all the Docs versions are knowledgeable about the number of available
versions and so on.
* Add support for injecting {{$page.currentVersion}} in markdown files
including code snippets, and links. For example: Assets versioning in code
snipper, Assets versioning in markdown IMG tags. Not the same but similar is
added to the Stylus:
https://github.com/handsontable/handsontable/blob/budnix-develop/docs/.vuepress/theme/components/ThemeSwitcher.vue#L146.
* Following the new Docs structure there are no such things as staging for the
previously released Docs. The staging exists only for the "next" version. In
the previous Docs structure, there was the possibility to push the changes to
the older Docs version and see the results in the staging environment before
merging. This can be treated as cons.