Allow to open cell's editor that is rendered out of the viewport (#9842)
The commit fixes the regression bug that prevents opening the cell editor when
it is rendered out of the table's viewport. On the selected cell, when we start
to enter text, the cell goes into edit mode. When the cell is editable - is not
read-only, with defined editor class or non-hidden - then the viewport is
scrolled to the cell that next goes into edit mode.

The commit extends the `openEditor` method for the 3rd `enableFullEditMode`
argument. The method is not available in the public API, even as a hidden
method in the Handsontable instance, so I guess the change can be released as
PATCH.