Fix mobile selection handlers to not extend the table's container (#9850)
The commit fixes an issue where the selection handlers could extend the table's
container causing the cells to desynchronize with headers. The issue happened
only when the cells on the table's edge were selected. To fix the issue, the
commit implements selection handlers positioning to the center of the table if
the edge cells are selected.