Some clarifications for the React docs (#9902)
* explain recommended way of declaring data

as it was discussed in https://github.com/handsontable/handsontable/pull/9787#discussion_r958395735

The reasons are:
1. the reader can use all other props as they are (by copying from the snippet), except for `data`, which for sure will be different in their app
2. `data` is the most likely to come from another level in the application and passed here using a variable

* replace attribute with property

because the examples clearly show the use of a property

* clarifications around the term "Handsontable hooks"

* attribute, property -> prop

Because React components don't have attributes or properties, they have props

Note that JSX has attributes, though. They are used to provide the values to props.