consistently use `undefined`, not `void`, in the API reference
Some new instances of `void` were added in PR #9724. Both word mean more or less the same, but we should stick to just using one of them to avoid confusion. 

What others think is the best word:
- JavaScript community is slightly skewed towards undefined: https://stackoverflow.com/questions/4759175/how-to-return-void-in-jsdoc
- In the ESLint rule, both void and undefined are acceptable: https://github.com/gajus/eslint-plugin-jsdoc

What we use: I counted 7 uses of `void` and 50+ uses of `undefined` in our JSDoc, so `undefined` is the way to go just to be consistent

What I think: `void` is an operator and `undefined` is a primitive data type and value. You cannot return an operator, you return a value. So it should be `undefined`.

In the TS definitions files, we usually write `undefined` in the function params and `void` in the return types. I agree that `void` is preferable in TypeScript type definition files because it doesn't require a `return` statement: https://stackoverflow.com/a/50574018/116087