Docs: Adding 12.1.3 docs  (#9913)
* fix typos and reword the changelog entries

* add the links to more info about the release

* add type definition changes that were skipped in the changelog

* change the order of information to the order of importance

* manually add typo fix from #9612

because the PR 9612 changed all versions of the docs except for "next"

* fix a typo in the name added in #9701

* correct the changes from #9701

that mixed the perspective of the developer and the end-user

* further clarifications that should have been added in #9701

* clarify that trimWhitespace affects both the renderer and the editor

the trimWhitespace was not presented in the docs before 12.1.3, due to an error fixed in #9872

* move the information added in PR #9718

to a place where it fits contextually

* consistently use `undefined`, not `void`, in the API reference

Some new instances of `void` were added in PR #9724. Both word mean more or less the same, but we should stick to just using one of them to avoid confusion. 

What others think is the best word:
- JavaScript community is slightly skewed towards undefined: https://stackoverflow.com/questions/4759175/how-to-return-void-in-jsdoc
- In the ESLint rule, both void and undefined are acceptable: https://github.com/gajus/eslint-plugin-jsdoc

What we use: I counted 7 uses of `void` and 50+ uses of `undefined` in our JSDoc, so `undefined` is the way to go just to be consistent

What I think: `void` is an operator and `undefined` is a primitive data type and value. You cannot return an operator, you return a value. So it should be `undefined`.

In the TS definitions files, we usually write `undefined` in the function params and `void` in the return types. I agree that `void` is preferable in TypeScript type definition files because it doesn't require a `return` statement: https://stackoverflow.com/a/50574018/116087

* fix typo added in #8468

* change the date of the release to the anticipated on

* copy the changelog to the release notes

* Update description for the getSelectedColumn Filters method

* update the versions of the mentioned dependencies according to package-lock.json

* update the versions of Node.js and npm

* change the explanation of the sorting algorithm change

* fix grammar

* improve the introduction

* implement a suggestion by @budnix

https://github.com/handsontable/handsontable/pull/9913/files#r976442755

Co-authored-by: Krzysztof Budnik <571316+budnix@users.noreply.github.com>