Implement ResizeObserver to the table's root element (#10086)
The commit fixes the issue related to the UI glitches that happened while
expanding Handsontable's root element from outside the component. Or expanding
the window size that is caused by the context menu or dropdown menu when the
table is initialized within an iframe.

The commit implements a more general solution based on the `ResizeObserver`
API. The fix mounts the `ResizeObserver` instance to the root Handsontable
element. For each size change, the `refreshDimensions` method is called. The
solution works almost the same as the implementation for window resize.

To ensure that Handsontable runs in the unit tests (Jest with `jsdom` env), the
polyfill (`resize-observer-polyfill`) has to be installed and imported to the
test runner setup files.

The commit also updates the `puppeteer` to the latest version.