Improve Docs examples (#10044)
The commit improves all wrapper code snippets so that they are 1:1 copyable to
the developer's app. To do that, the commit implements two expressions as code
comments that do the following things:
1. `/* start:skip-in-preview */` and `/* end:skip-in-preview */` - Code placed
between the expressions is not visible in the "Preview" tab but is still
transpiled and runs on the Docs side, and it's visible in the code tab in the
jsFiddle (in the JS code tab);
2. `/* start:skip-in-compilation */` and `/* end:skip-in-compilation */` - Code
placed between the expressions is ignored, so it does not run on the Docs side.
The code is also not included in the jsFiddle's JS tab. The code is only
visible in the "Preview" tab as part of some framework requirements that hard
to emulate on the Docs side. The comments expressions are stripped out.

Moreover, the PR introduces the following miscellaneous things:
* Update all JavaScript examples. Add missing imports (import Handsontable from
`'handsontable'`; and `import 'handsontable/dist/handsontable.full.min.css'`);
* All `Handsontable.dom` private API invocations are removed. The DOM helper is
not available in the modularized Handsontable examples version.
* All `Handsontable.helpers` private API invocations are removed from demos as
they are unavailable in the modularized Handsontable (`handsontable/base`)
version. The native JS API function equivalents are used. The helper used for
creating the dataset (`createSpreadsheetData`) is replaced with a dataset
pasted directly to the code, and for bigger datasets, a basic snippet with a
multidimensional array generator is used.