Change the base branch the Docs prod branch is created from (#10122)
The commit improves the creation of the Docs production branches after the
Handsontable release process. Previously the branch was created from the
default (develop) branch, which may cause situations where non-released
features or code changes can be deployed with an already published version. The
commit eliminates that problem by updating the script, which creates the Docs
branch from the git tag of the released HoT version.