Fix an issue where typing in editors causes the IME editor to disappear (#10146)
The commit fixes an issue where typing in some editors like dropdown causes the
IME to disappear, resulting in typing the wrong letter. The fix is caused by
#6916, where the `blur()` is called every time the cell is selected. The
solution to the problem is adding logic that blurs the active element only when
the previously active element was not an element that belongs to the
Handsontable. Thanks to that, the fix for #6916 is still valid, and the IME
does not disappear.