Fix cell meta objects desync when declared in the `afterCreate{Row,Column}` hooks (#10136)
The commit fixes the issue of cell meta objects desynchronization with the
dataset. The bug happens when the new cell meta objects are modified in the
afterCreateRow or afterCreateCol hooks.

The main thing that caused the bug was the fact that the `MetaManager` was
called too late, after the afterCreate{Row,Col} hook was triggered. After
changing the index value (changed in #9955) to the index of the inserted
row/column, adding cell meta objects in the hook causes double cells shifting.

It's important to keep the `MetaManager` and `DataMap` in sync that's why I
decided to move the `MetaManager` instance to the `DataMap`. The
synchronization happens right after the dataset changes but before the
`afterCreate...` hook.