Improve triggering the Visual Tests job (#10228)
The commit adjusts the Visual Tests job that should run only when one of the
ancestor tasks finishes its job with the status `success`. Moreover, it turned
out that there were some cases where the Visual Tests job failed as it tried to
extract Handsontable artifacts that did not exist. The commit fixes that bug
too.