Implement keyboard navigation in headers (#10301)
The PR adds support for keyboard navigation in headers (including top and
bottom corners) using currently supported keyboard shortcuts e.g., Tab (next
cell/header), Shift+Tab (previous cell/header), Arrow keys and more.

Major (breaking change)
* Changed the signature of the CellCoords:isValid() and CellRange:isValid()
methods. The method in the origin state was inconvenient to use as it requires
passing a whole Walkontable instance which causes it to be even more difficult
to use, especially for 3rd app developers;
* Change how the CellRange:isSingle() method works. With the new navigable
headers feature, the method returns true not only when the single cell is
selected but also when the single header is selected;

Minor
* Add a new option navigableHeaders;
* Add new isSingleCell(), isSingleHeader () and containsHeaders() methods to
the CellRange class;
* Extend the Core methods such as selectColumns(), selectRows(), and
selectAll() for the ability to change the position of the cell focus
highlight;

Miscellaneous
* Reorganize the shortcuts and collect them from different parts of the code
(Core and EditorManager) and move them to the separate module called
shortcutContexts;
* Add an internal helper called clamp that allows easily apply constraints to
the number;
* Extend the E2E AsciiTable helper for support for nested headers and navigable
headers features;
* Add a new toEqualCellRange()Jasmine matcher that allows easily test of the
selection. For example, instead of writing

Supported keyboard shortcuts in headers
* Tab navigation (Tab, Shift+Tab);
* Arrow keys navigation;
* Single-cell navigation to the most left/right position (Home, End);
* Selecting all cells and headers (Cmd/Ctrl+A);
* Expanding selection when columns or rows are selected (Shift+Arrow keys);
* Expanding selection to the most left/right/top/bottom position when columns
or rows are selected (Shift+Cmd/Ctrl+Arrow keys);

Co-authored-by: Jakub Wiśniewski <jakub.wisniewski@handsontable.com>