Add support for keyboard shortcuts that allow sorting columns (#10410)
The PR adds support for new keyboard shortcuts that allows sorting the
column/multiple columns based on the currently selected column header. The
keyboard shortcuts are active only when the header is focused.

Expected working new keyboard shortcuts
* Enter Sorts the column (if the ColumnSorting or MultipleColumnSorting plugins
are enabled) in a specific order: Ascending > Descending > Restore original
order. If there is initialized the multiple columns order, the whole state is
cleared, and only the selected column is sorted;
* Ctrl/Cmd+Enter Sorts the column in the same order as above, and it works only
if the MultipleColumnSorting plugin is enabled. Sorting other columns does not
clear the previous columns sorting state but adds new sorting order to the
stack.