Add support for keyboard shortcut that allows opening the context menu (#10417)
The commit adds support for a new keyboard shortcut that allows opening the
context menu (from ContextMenu plugin). The keyboard shortcut is active for
cells and for headers as well.

Expected working new keyboard shortcuts
* `Cmd/Ctrl+Shift+\` or `Shift+F10` Opens the context menu. The menu is
rendered at the bottom of the focused cell by default. If there is not free
space in the viewport to render the menu it tries to render it on the other
side of the cell.