Add support for keyboard shortcut that allows clearing the filters (#10420)
The commit adds support for a new keyboard shortcut that allows clearing the
state of the filters. The shortcut is active only when the Filters plugin is
enabled.

Expected working new keyboard shortcuts
* Alt/Option+A Clears the filters from all columns. The shortcut is active for
the whole table. There is not necessary to select a column or a cell to clear
the states. Clearing the states should keep the position of the current
selection the same.