Add the ability to traverse the grid using TAB key throughout the page (#10430)
The PR adds the ability to traverse the grid using Tab and Shift+Tab throughout
the page keeping the correct order of the native browser focus. This is a
breaking change as from now on when the last cell is selected (or the last cell
in a row), pressing the Tab key deactivates the table (the table no more
listens for any keyboard shortcuts) and moves the browser's focus to the next
element.

The PR introduces a new API:

* The new table option disableTabNavigation (default false). The option, when
set as true, disables the table navigation. In that mode, the arrow keys are
supposed to use to navigate the table. Pressing the Tab or Shift+Tab moves the
browser's focus to the next or previous element on the page. When set as false,
there is a possibility to navigate the table, but when the cell selection
reaches the last cell (or the last cell in a row when autoWrapRow: false) then
the table deactivates itself, and the browser's focus is moved to the next
element on the page;
* The new modifyFocusOnTabNavigation hook. The hook allows changing the cell
selection after the table is activated by the browser's focus triggered by Tab
or Shift+Tab keys. By default, the first visible cell (or header) is selected
when the table is activated from the above element. When it's activated from
the element below, the last cell is selected;
* The Core listen method is improved for pages where multiple Handsontable
instances are created. From now on, when one of the instances is activated, the
other instance that the activation was taken from calls 'unlisten' hook on its
instance;

Co-authored-by: Wojciech Szymański <141330+wszymanski@users.noreply.github.com>